refactor download package structure to improve consistency
