+++ annotation-rest/src/main/java/uk/ac/ebi/quickgo/annotation/download/http/StopStreamException.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.quickgo.annotation.service.http;
+package uk.ac.ebi.quickgo.annotation.download.http;
 
 import java.util.stream.Stream;
