+++ annotation-rest/src/test/java/uk/ac/ebi/quickgo/annotation/download/DownloadTaskExecutorReadPropertiesTest.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.quickgo.annotation.service.download;
+package uk.ac.ebi.quickgo.annotation.download;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,9 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -15,43 +17,43 @@ import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static uk.ac.ebi.quickgo.annotation.service.download.TaskExecutorProperties.*;
 
 /**
- * Check default properties for configuring the download {@link ThreadPoolTaskExecutor}, were used.
+ * Check properties for configuring the download {@link ThreadPoolTaskExecutor}, were read and used.
  *
  * Created 23/01/17
  * @author Edd
  */
+@ActiveProfiles("download-task-executor-test")
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = DownloadTaskExecutorDefaultPropertiesTest.FakeApplication.class)
-public class DownloadTaskExecutorDefaultPropertiesTest {
+@SpringApplicationConfiguration(classes = DownloadTaskExecutorReadPropertiesTest.FakeApplication.class)
+public class DownloadTaskExecutorReadPropertiesTest {
     @Autowired
     private ThreadPoolTaskExecutor taskExecutor;
 
     @Test
     public void corePoolSizeIsPopulated() {
-        assertThat(taskExecutor.getCorePoolSize(), is(DEFAULT_CORE_POOL_SIZE));
+        assertThat(taskExecutor.getCorePoolSize(), is(10));
     }
 
     @Test
     public void maxPoolSizeIsPopulated() {
-        assertThat(taskExecutor.getMaxPoolSize(), is(MAX_POOL_SIZE));
+        assertThat(taskExecutor.getMaxPoolSize(), is(20));
     }
 
     @Test
     public void queueCapacityIsPopulated() {
-        assertThat(taskExecutor.getThreadPoolExecutor().getQueue().remainingCapacity(), is(QUEUE_CAPACITY));
+        assertThat(taskExecutor.getThreadPoolExecutor().getQueue().remainingCapacity(), is(100));
     }
 
     @Test
     public void keepAliveIsPopulated() {
-        assertThat(taskExecutor.getKeepAliveSeconds(), is(KEEP_ALIVE_SECONDS));
+        assertThat(taskExecutor.getKeepAliveSeconds(), is(600));
     }
 
     @Test
     public void allowCoreThreadTimeoutIsPopulated() {
-        assertThat(taskExecutor.getThreadPoolExecutor().allowsCoreThreadTimeOut(), is(ALLOW_CORE_THREAD_TIMEOUT));
+        assertThat(taskExecutor.getThreadPoolExecutor().allowsCoreThreadTimeOut(), is(false));
     }
 
     /**
@@ -60,14 +62,19 @@ public class DownloadTaskExecutorDefaultPropertiesTest {
      */
     @Test
     public void verifySetWaitForTasksToCompleteOnShutdownIsCalled() {
+        TaskExecutorProperties properties = new TaskExecutorProperties();
+        boolean expectedWaitForTasksToComplete = true;
+        properties.setWaitForTasksToCompleteOnShutdown(expectedWaitForTasksToComplete);
         ThreadPoolTaskExecutor taskExecutor = mock(ThreadPoolTaskExecutor.class);
 
         DownloadConfig config = new DownloadConfig();
+        config.setTaskExecutor(properties);
         config.taskExecutor(taskExecutor);
 
-        verify(taskExecutor, times(1)).setWaitForTasksToCompleteOnShutdown(WAIT_FOR_TASKS_TO_COMPLETE_ON_SHUTDOWN);
+        verify(taskExecutor, times(1)).setWaitForTasksToCompleteOnShutdown(expectedWaitForTasksToComplete);
     }
 
+    @Profile("download-task-executor-test")
     @Configuration
     @EnableAutoConfiguration
     @Import(DownloadConfig.class)
